+++ nexus/checks.py
@@ -12,7 +12,7 @@ def check_requirements(app_configs, **kwargs):
     if getattr(settings, 'NEXUS_SKIP_INSTALLED_APPS_REQUIREMENTS', False):
         reqs = ()
     else:
-        reqs = ('django.contrib.auth', 'django.contrib.sessions')
+        reqs = ('django.contrib.admin', 'django.contrib.auth', 'django.contrib.sessions')
 
     for req in reqs:
         if req not in settings.INSTALLED_APPS: