+++ nexus/conf.py
@@ -5,7 +5,6 @@ class Settings(object):
     """
     Shadow Django's settings with a little logic
     """
-
     @property
     def MEDIA_PREFIX(self):
         prefix = getattr(settings, 'NEXUS_MEDIA_PREFIX', '/nexus/media/')